requirements.txt
fastapi
uvicorn
streamlit
sentence-transformers
faiss-cpu
beautifulsoup4
requests
pydantic
